requirements.txt
google-genai
python-dotenv
Flask
groq
